requirements.txt
Django==1.4
Markdown==2.2.0
PIL==1.1.7
South==0.7.6
argparse==1.2.1
-e git://github.com/earle/django-bootstrap.git@9ec5697aee8fed5b612ac8a0a40bb1921e9390e8#egg=django-bootstrap-dev
cmsplugin-blog==1.1.2
django-appconf==0.5
django-classy-tags==0.3.4.1
django-cms==2.3
django-debug-toolbar==0.9.4
django-markitup==1.0.0
django-missing==0.1.5
django-mptt==0.5.2
django-sekizai==0.6.1
django-staticfiles==1.2.1
django-tagging==0.3.1
djangocms-utils==0.9.5
html5lib==0.95
icalendar==3.0.1b2
psycopg2==2.4.5
pytz==2012d
simple-translation==0.8.6
-e git://github.com/sophilabs/symposion.git@8af8a4ab46736bc32f5e8acbbfcda4d8232953a1#egg=symposion-dev
wsgiref==0.1.2
django-email-confirmation==0.2
Pinax==0.9a2
django-timezones==0.2